requirements.txt
paho-mqtt
cryptography
meshtastic
dotenv
sqlalchemy
pydantic
pynacl
prettytable